home improve

+++ application/models/m_upload.php
@@ -116,44 +116,49 @@ Class m_upload extends CI_Model {
 
     public function upload_multi_file($id, $input_name, $file_id = NULL) {
         $this->load->library('upload');
-
-        $_FILES = $this->multifile($_FILES[$input_name]);
         $i = 0;
+        $_FILES = $this->multifile($_FILES['file']);
         foreach ($_FILES as $file => $file_data) {
-            $this->upload->initialize($this->set_upload_file());
-            $this->upload->do_upload($file);
+            // No problems with the file
+            if ($file_data['error'] == 0) {
+                $this->upload->initialize($this->set_upload_file());
+                // So lets upload
+                if ($this->upload->do_upload($file)) {
+                    $finfo = $this->upload->data();
+                    //insert to database
+                    $data_file = array(
+                        'file_name' => $finfo['file_name'],
+                        'file_path' => $finfo['file_path'],
+                        'file_full_path' => $finfo['full_path'],
+                    );
 
-            $finfo = $this->upload->data();
-            //insert to database
-            $data_file = array(
-                'file_name' => $finfo['file_name'],
-                'file_path' => $finfo['file_path'],
-                'file_full_path' => $finfo['full_path'],
-            );
-            if ($file_id == NULL) {
-                $this->db->trans_start();
-                $this->db->insert('files', $data_file);
-                $file_id = $this->db->insert_id();
-                $this->db->trans_complete();
-                $t = $this->input->post('txtTitle');
-                if ($t[$i] == NULL || $t[$i] == '') {
-                    $title = $finfo['orig_name'];
-                } else {
-                    $title = $t[$i];
-                }
-                $f = array(
-                    'news_id' => $id,
-                    'file_id' => $file_id,
-                    'title' => $title,
-                );
-                $this->db->insert('news_has_files', $f);
-            } else {
-                $this->deleteFile($file_id);
-                $this->db->where('file_id', $file_id);
-                $this->db->update('files', $data_file);
+                    if ($file_id == NULL) {
+                        $this->db->trans_start();
+                        $this->db->insert('files', $data_file);
+                        $file_id = $this->db->insert_id();
+                        $this->db->trans_complete();
+                        $t = $this->input->post('txtTitle');
+                        if ($t[$i] == NULL || $t[$i] == '') {
+                            $title = $finfo['orig_name'];
+                        } else {
+                            $title = $t[$i];
+                        }
+                        $f = array(
+                            'news_id' => $id,
+                            'file_id' => $file_id,
+                            'title' => $title,
+                        );
+                        $this->db->insert('news_has_files', $f);
+                    } else {
+                        $this->deleteFile($file_id);
+                        $this->db->where('file_id', $file_id);
+                        $this->db->update('files', $data_file);
 //                        return $file_id;
+                    }
+                } else {
+                    $errors = $this->upload->display_errors();
+                }
             }
-
             $i++;
         }
     }
@@ -174,24 +179,22 @@ Class m_upload extends CI_Model {
 // Codeigniter Upload Multiple File
     public function multifile($filedata) { // $_FILES['files'];
         if (count($filedata) == 0)
-            return NULL;
+            return FALSE;
 
         $files = array();
         $all_files = $filedata['name'];
         $i = 0;
 
         foreach ($all_files as $filename) {
-            if (!empty($files[++$i]['name'])) {
-                $files[$i]['name'] = $filename;
-                $files[$i]['type'] = current($filedata['type']);
-                next($filedata['type']);
-                $files[$i]['tmp_name'] = current($filedata['tmp_name']);
-                next($filedata['tmp_name']);
-                $files[$i]['error'] = current($filedata['error']);
-                next($filedata['error']);
-                $files[$i]['size'] = current($filedata['size']);
-                next($filedata['size']);
-            }
+            $files[++$i]['name'] = $filename;
+            $files[$i]['type'] = current($filedata['type']);
+            next($filedata['type']);
+            $files[$i]['tmp_name'] = current($filedata['tmp_name']);
+            next($filedata['tmp_name']);
+            $files[$i]['error'] = current($filedata['error']);
+            next($filedata['error']);
+            $files[$i]['size'] = current($filedata['size']);
+            next($filedata['size']);
         }
 
         return $files;
@@ -208,3 +211,43 @@ Class m_upload extends CI_Model {
     }
 
 }
+
+//  $_FILES = $this->multifile($_FILES[$input_name]);
+//        $i = 0;
+//        foreach ($_FILES as $file => $file_data) {
+//            $this->upload->initialize($this->set_upload_file());
+//            $this->upload->do_upload($file);
+//
+//            $finfo = $this->upload->data();
+//            //insert to database
+//            $data_file = array(
+//                'file_name' => $finfo['file_name'],
+//                'file_path' => $finfo['file_path'],
+//                'file_full_path' => $finfo['full_path'],
+//            );
+//            if ($file_id == NULL) {
+//                $this->db->trans_start();
+//                $this->db->insert('files', $data_file);
+//                $file_id = $this->db->insert_id();
+//                $this->db->trans_complete();
+//                $t = $this->input->post('txtTitle');
+//                if ($t[$i] == NULL || $t[$i] == '') {
+//                    $title = $finfo['orig_name'];
+//                } else {
+//                    $title = $t[$i];
+//                }
+//                $f = array(
+//                    'news_id' => $id,
+//                    'file_id' => $file_id,
+//                    'title' => $title,
+//                );
+//                $this->db->insert('news_has_files', $f);
+//            } else {
+//                $this->deleteFile($file_id);
+//                $this->db->where('file_id', $file_id);
+//                $this->db->update('files', $data_file);
+////                        return $file_id;
+//            }
+//
+//            $i++;
+//        }
